Initialize currency data on server startup
