+++ zlotowka-backend/src/main/java/com/agh/zlotowka/controller/UserController.java
@@ -1,32 +1,24 @@
 package com.agh.zlotowka.controller;
 
-import com.agh.zlotowka.model.User;
 import com.agh.zlotowka.dto.UserResponse;
-import com.agh.zlotowka.repository.UserRepository;
 import com.agh.zlotowka.security.CustomUserDetails;
 import com.agh.zlotowka.service.UserService;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/user")
 public class UserController {
 
-    private final UserRepository userRepository;
     private final UserService userService;
 
     @GetMapping("/account")
     public ResponseEntity<UserResponse> getCurrentUser(@AuthenticationPrincipal CustomUserDetails userDetails) {
         return ResponseEntity.ok(userService.getCurrentUser(userDetails));
     }
-    @PostMapping("/currency")
-    public ResponseEntity<Void> addCurrencies() {
-        userService.addCurrencies();
-        return ResponseEntity.noContent().build();
-    }
-
 }
